Break objective get_* utils to their own file
