+++ desc/objectives/__init__.py
@@ -30,6 +30,13 @@ from ._qs import (
     QuasisymmetryTwoTerm,
 )
 from ._stability import MagneticWell, MercierStability
+from .getters import (
+    get_equilibrium_objective,
+    get_fixed_axis_constraints,
+    get_fixed_boundary_constraints,
+    get_NAE_constraints,
+    maybe_add_self_consistency,
+)
 from .linear_objectives import (
     AxisRSelfConsistency,
     AxisZSelfConsistency,
@@ -55,9 +62,3 @@ from .linear_objectives import (
     FixThetaSFL,
 )
 from .objective_funs import ObjectiveFunction
-from .utils import (
-    get_equilibrium_objective,
-    get_fixed_axis_constraints,
-    get_fixed_boundary_constraints,
-    get_NAE_constraints,
-)
